+++ src/constants/Command.ts
@@ -20,6 +20,7 @@ export const COMMAND = {
   stepMoveUp: `${EXTENSION_NAME}.stepMoveUp`,
   stepMoveDown: `${EXTENSION_NAME}.stepMoveDown`,
   viewStep: `${EXTENSION_NAME}.viewStep`,
+  createSnapshot: `${EXTENSION_NAME}.createSnapshot`,
   // Output
   showOutputChannel: `${EXTENSION_NAME}.showOutputChannel`,
   // Panel view