+++ src/constants/General.ts
@@ -1,4 +1,6 @@
 export const General = {
   demoFolder: ".demo",
   variablesFile: "variables.json",
+  snapshotsFolder: "snapshots",
+  patchesFolder: "patches",
 };